Core/Unit: forgotten explicit cast

+++ src/server/game/Server/Protocol/Handlers/MiscHandler.cpp
@@ -1079,7 +1079,7 @@ void WorldSession::HandleObjectUpdateFailed(WorldPacket& recv_data)
     recv_data.ReadByteSeq(guid[5]);
 
     WorldObject* obj = ObjectAccessor::GetWorldObject(*GetPlayer(), guid);
-    sLog->outError("Object update failed for object guid %u (" UI64FMTD ") (%s), player %s", GUID_LOPART(uint64(guid)), guid, obj?obj->GetName():"not-found", GetPlayer()->GetName());
+    sLog->outError("Object update failed for object guid %u (" UI64FMTD ") (%s), player %s", GUID_LOPART(uint64(guid)), uint64(guid), obj?obj->GetName():"not-found", GetPlayer()->GetName());
 }
 
 void WorldSession::HandleSetActionButtonOpcode(WorldPacket& recv_data)
